# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Logo Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Logo product operations
  - name: categories
    description: Category operations
  - name: orders
    description: Order operations
  - name: wishlist
    description: Wishlist operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [popular, newest, price_asc, price_desc]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user's orders (requires auth)
      responses:
        "200":
          description: User's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order (requires auth)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user's wishlist (requires auth)
      responses:
        "200":
          description: User's wishlist products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
        "401":
          description: Unauthorized

  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist (requires auth)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from wishlist
        "401":
          description: Unauthorized


  /custom-logo:
    post:
      operationId: createCustomLogo
      tags: [custom-logo]
      summary: Submit a custom logo design request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomLogoRequestInput'
      responses:
        '201':
          description: Request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  id:
                    type: string
        '400':
          description: Validation error
        '500':
          description: Server error

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        style:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        previewUrls:
          type: array
          items:
            type: string
        featured:
          type: boolean
        popularity:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - price
        - category
        - style
        - tags
        - imageUrl
        - previewUrls
        - featured
        - popularity
        - createdAt

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        iconName:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - iconName
        - productCount

    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required:
        - id
        - userId
        - status
        - totalAmount
        - items
        - createdAt

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        price:
          type: number
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - orderId
        - productId
        - price
        - product

    CreateOrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        utr:
          type: string
        screenshotUrl:
          type: string
      required:
        - items

    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        price:
          type: number
      required:
        - productId
        - price

    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        productId:
          type: integer
      required:
        - id
        - userId
        - productId

    WishlistInput:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId


    CustomLogoRequestInput:
      type: object
      properties:
        businessName:
          type: string
        slogan:
          type: string
        description:
          type: string
        industry:
          type: string
        targetAudience:
          type: string
        top3Things:
          type: array
          items:
            type: string
        ideas:
          type: string
        colors:
          type: string
        styles:
          type: array
          items:
            type: string
        sliders:
          type: object
          properties:
            feminineMasculine:
              type: number
            simpleComplex:
              type: number
            grayColorful:
              type: number
            subtleBright:
              type: number
            quietLoud:
              type: number
            necessityLuxury:
              type: number
            expensiveEconomical:
              type: number
            playfulSerious:
              type: number
            modernClassic:
              type: number
            rawRefined:
              type: number
            exoticCommonplace:
              type: number
            sportyElegant:
              type: number
            adventureSecure:
              type: number
        paymentId:
          type: string
        uploadLinks:
          type: array
          items:
            type: string
      required:
        - businessName
        - description
        - industry
        - targetAudience
        - top3Things
        - colors
        - styles
        - sliders
        - paymentId