# lib/api-zod/src/generated/types/customLogoRequestInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logo Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomLogoRequestInputSliders } from "./customLogoRequestInputSliders";

export interface CustomLogoRequestInput {
  businessName: string;
  slogan?: string;
  description: string;
  industry: string;
  targetAudience: string;
  top3Things: string[];
  ideas?: string;
  colors: string;
  styles: string[];
  sliders: CustomLogoRequestInputSliders;
  paymentId: string;
  uploadLinks?: string[];
}